chore: Template 사용 유무에 따라 파일명 다르게 수정
