+++ lib/Model/Folder/FolderModel.dart
@@ -1,12 +1,12 @@
 // FolderModel.dart
-import '../Problem/ProblemModel.dart';
+import '../Problem/ProblemModelWithTemplate.dart';
 import 'FolderThumbnailModel.dart';
 
 class FolderModel {
   final int folderId;
   final String folderName;
   final FolderThumbnailModel? parentFolder;
-  final List<ProblemModel> problems;
+  final List<ProblemModelWithTemplate> problems;
   final List<FolderThumbnailModel> subFolderList;
   final DateTime? createdAt;
   final DateTime? updateAt;
@@ -30,7 +30,8 @@ class FolderModel {
     // problemList → problems
     final problemList = json['problemList'] as List<dynamic>? ?? [];
     final problems = problemList
-        .map((e) => ProblemModel.fromJson(e as Map<String, dynamic>))
+        .map(
+            (e) => ProblemModelWithTemplate.fromJson(e as Map<String, dynamic>))
         .toList();
 
     // subFolderList: List<Map> → List<FolderThumbnailModel>
