+++ lib/Model/Problem/ProblemThumbnailModel.dart
@@ -1,20 +1,20 @@
 class ProblemThumbnailModel {
   final int problemId;
   final String reference;
-  final String processImageUrl;
+  final String problemImageUrl;
   final DateTime? createdAt;
 
   ProblemThumbnailModel(
       {required this.problemId,
       required this.reference,
-      required this.processImageUrl,
+      required this.problemImageUrl,
       required this.createdAt});
 
   factory ProblemThumbnailModel.fromJson(Map<String, dynamic> json) {
     return ProblemThumbnailModel(
       problemId: json['problemId'],
       reference: json['reference'],
-      processImageUrl: json['processImageUrl'],
+      problemImageUrl: json['problemImageUrl'],
       createdAt: json['createdAt'],
     );
   }
