+++ lib/Screen/ProblemRegister/ProblemRegisterScreen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:ono/Model/Problem/ProblemModel.dart';
+import 'package:ono/Model/Problem/ProblemModelWithTemplate.dart';
 import 'package:provider/provider.dart';
 
 import '../../Module/Text/StandardText.dart';
@@ -7,7 +7,7 @@ import '../../Module/Theme/ThemeHandler.dart';
 import 'ProblemRegisterTemplateV2.dart';
 
 class ProblemRegisterScreenV2 extends StatefulWidget {
-  final ProblemModel? problemModel;
+  final ProblemModelWithTemplate? problemModel;
   final bool isEditMode;
 
   const ProblemRegisterScreenV2({
